+++ training_log_app/src/repositories/log_entry_repository.py
@@ -19,61 +19,72 @@ class LogEntryRepository:
 
         cursor = self.database.cursor()
 
-        entry_list = cursor.execute('''select * from Log_entries where user_id=? order by id desc''', [user.id]).fetchone()
+        entry_list = cursor.execute('''select * from Log_entries where
+                                 user_id=? order by id desc''', [user.id]).fetchone()
 
         self.database.commit()
 
-        return entry_list  # untested
-    
+        return entry_list
+
     def get_total_time_spent_by_user(self, user: User):
         '''Returns total time (sum of durations)
         user has spent training.
-        
+
         Arg: user - which user
-        
+
         Return: sum of durations'''
 
         cursor = self.database.cursor()
 
-        time = cursor.execute('''select ifnull(sum(duration),0) from Log_entries where user_id=?''', [user.id]).fetchone()
+        time = cursor.execute('''select ifnull(sum(duration),0)
+                                from Log_entries where user_id=?''',
+                              [user.id]).fetchone()
 
         return time[0]
-
 
     def get_amount_of_goals_achieved_by_user(self, user: User):
         '''Returns a tuple containing (a, b)
         where a represents how many times user
         has achieved a set goal and b represents
         amount of all entries by user.
-        
+
         Arg: user - which user
-        
+
         Return: specified tuple'''
 
         cursor = self.database.cursor()
 
-        goals_achieved = cursor.execute('''select ifnull(count(*),0) from Log_entries where user_id=? and was_last_goal_achieved=1''', [user.id]).fetchone()[0]
+        goals_achieved = cursor.execute('''select ifnull(count(*),0)
+                                        from Log_entries where user_id=?
+                                        and was_last_goal_achieved=1''',
+                                        [user.id]).fetchone()[0]
 
-        amount_of_entries = cursor.execute('''select ifnull(count(*),0) from Log_entries where user_id=?''', [user.id]).fetchone()[0]
+        amount_of_entries = cursor.execute('''select ifnull(count(*),0)
+                                        from Log_entries where user_id=?''',
+                                           [user.id]).fetchone()[0]
 
         self.database.commit()
 
-        return goals_achieved, amount_of_entries # untested
+        return goals_achieved, amount_of_entries
 
     def get_last_entry_with_user(self, user: User):
         '''Return users last entry.
         Args: user - which user
-        
+
         Returns: Users last entry. Or if no
         last entry, then text saying so.'''
 
         cursor = self.database.cursor()
 
-        entry = cursor.execute('''select * from Log_entries where user_id=? order by id desc limit 1''', [user.id]).fetchall()
+        entry = cursor.execute('''select * from Log_entries
+                                where user_id=? 
+                                order by id desc limit 1''',
+                               [user.id]).fetchall()
 
         self.database.commit()
 
-        return 'No previous entries' if entry is None else list(entry[0])
+        # if no entries a completely different frame needs to be run
+        return 'No previous entries' if len(entry) == 0 else list(entry[0])
 
     def create_entry(self, log_entry: LogEntry):
         '''Create entry to database.